BE/requirement.txt
quart
quart-cors
asyncio
numpy
torch
nistrng
pycryptodome